Move Vector2String and Color4String methods into Utils so other parts of application can access
